refactor: de-couple state transition validation from store and DB
