+++ packages/server-wallet/src/__test__/evm-validator.test.ts
@@ -2,24 +2,12 @@ import {utils} from 'ethers';
 import {toNitroState} from '@statechannels/wallet-core';
 
 import {appBytecode, COUNTING_APP_DEFINITION} from '../models/__test__/fixtures/app-bytecode';
-import {AppBytecode} from '../models/app-bytecode';
 import {createState} from '../wallet/__test__/fixtures/states';
 import {validateTransitionWithEVM} from '../evm-validator';
-import {defaultConfig} from '../config';
-import {testKnex as knex} from '../../jest/knex-setup-teardown';
 
 const UNDEFINED_APP_DEFINITION = '0x88c26ec40DC653973C599A1a0762678e795F879F';
 
-beforeEach(async () => {
-  await knex('app_bytecode').truncate();
-  await AppBytecode.query(knex).insert([appBytecode()]);
-});
-
 it('returns true for a valid transition', async () => {
-  // Sanity check that we're validating with byte code
-  expect(
-    await AppBytecode.getBytecode(defaultConfig.chainNetworkID, COUNTING_APP_DEFINITION, knex)
-  ).toBeDefined();
   const fromState = toNitroState(
     createState({
       appDefinition: COUNTING_APP_DEFINITION,
@@ -32,14 +20,10 @@ it('returns true for a valid transition', async () => {
       appData: utils.defaultAbiCoder.encode(['uint256'], [2]),
     })
   );
-  expect(await validateTransitionWithEVM(fromState, toState, knex as any)).toBe(true);
+  expect(await validateTransitionWithEVM(fromState, toState, appBytecode().appBytecode)).toBe(true);
 });
 
 it('returns false for an invalid transition', async () => {
-  // Sanity check that we're validating with byte code
-  expect(
-    await AppBytecode.getBytecode(defaultConfig.chainNetworkID, COUNTING_APP_DEFINITION, knex)
-  ).toBeDefined();
   const fromState = toNitroState(
     createState({
       appDefinition: COUNTING_APP_DEFINITION,
@@ -52,18 +36,16 @@ it('returns false for an invalid transition', async () => {
       appData: utils.defaultAbiCoder.encode(['uint256'], [1]),
     })
   );
-  expect(await validateTransitionWithEVM(fromState, toState, knex as any)).toBe(false);
+  expect(await validateTransitionWithEVM(fromState, toState, appBytecode().appBytecode)).toBe(
+    false
+  );
 });
 
 it('returns false when no byte code exists for the app definition', async () => {
-  // Sanity check that the bytecode doesn't exist
-  expect(
-    await AppBytecode.getBytecode(defaultConfig.chainNetworkID, UNDEFINED_APP_DEFINITION, knex)
-  ).toBeUndefined();
   const state = toNitroState(
     createState({
       appDefinition: UNDEFINED_APP_DEFINITION,
     })
   );
-  expect(await validateTransitionWithEVM(state, state, knex as any)).toBe(false);
+  expect(await validateTransitionWithEVM(state, state, '0x')).toBe(false);
 });
